+++ x/pricefeed/client/cli/gen_whitelist_oracle_test.go
@@ -5,12 +5,10 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/NibiruChain/nibiru/x/testutil"
-
 	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/server"
 	"github.com/cosmos/cosmos-sdk/simapp"
+	sdktestutil "github.com/cosmos/cosmos-sdk/testutil"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	"github.com/cosmos/cosmos-sdk/x/genutil"
 	genutiltest "github.com/cosmos/cosmos-sdk/x/genutil/client/testutil"
@@ -19,56 +17,18 @@ import (
 	"github.com/tendermint/tendermint/libs/log"
 
 	"github.com/NibiruChain/nibiru/x/pricefeed/client/cli"
+	"github.com/NibiruChain/nibiru/x/testutil"
 )
 
 var testModuleBasicManager = module.NewBasicManager(genutil.AppModuleBasic{})
 
 // Tests "add-genesis-oracle", a command that adds a oracle address to genesis.json
 func TestAddGenesisWhitelistOracleCmd(t *testing.T) {
-	type TestCase struct {
+	testCases := []struct {
 		name        string
 		oracles     string
 		expectError bool
-	}
-
-	var executeTest = func(t *testing.T, testCase TestCase) {
-		tc := testCase
-		t.Run(tc.name, func(t *testing.T) {
-			home := t.TempDir()
-			logger := log.NewNopLogger()
-			cfg, err := genutiltest.CreateDefaultTendermintConfig(home)
-			require.NoError(t, err)
-
-			appCodec := simapp.MakeTestEncodingConfig().Marshaler
-			err = genutiltest.ExecInitCmd(
-				testModuleBasicManager, home, appCodec)
-			require.NoError(t, err)
-
-			serverCtx := server.NewContext(viper.New(), cfg, logger)
-			clientCtx := client.Context{}.WithCodec(appCodec).WithHomeDir(home)
-
-			ctx := context.Background()
-			ctx = context.WithValue(ctx, client.ClientContextKey, &clientCtx)
-			ctx = context.WithValue(ctx, server.ServerContextKey, serverCtx)
-
-			cmd := cli.AddWhitelistGenesisOracle("home")
-			cmd.SetArgs(
-				[]string{
-					tc.oracles,
-					fmt.Sprintf("--%s=home", flags.FlagHome),
-				},
-			)
-
-			if tc.expectError {
-				err := cmd.ExecuteContext(ctx)
-				require.Error(t, err)
-			} else {
-				require.NoError(t, cmd.ExecuteContext(ctx))
-			}
-		})
-	}
-
-	testCases := []TestCase{
+	}{
 		{
 			name:        "add single oracle",
 			oracles:     testutil.AccAddress().String(),
@@ -96,7 +56,34 @@ func TestAddGenesisWhitelistOracleCmd(t *testing.T) {
 		},
 	}
 
-	for _, testCase := range testCases {
-		executeTest(t, testCase)
+	for _, tc := range testCases {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			home := t.TempDir()
+			logger := log.NewNopLogger()
+			cfg, err := genutiltest.CreateDefaultTendermintConfig(home)
+			require.NoError(t, err)
+
+			appCodec := simapp.MakeTestEncodingConfig().Marshaler
+			err = genutiltest.ExecInitCmd(testModuleBasicManager, home, appCodec)
+			require.NoError(t, err)
+
+			serverCtx := server.NewContext(viper.New(), cfg, logger)
+
+			cmd := cli.AddWhitelistGenesisOracle(home)
+			cmd.SetArgs([]string{tc.oracles})
+			_, out := sdktestutil.ApplyMockIO(cmd)
+			clientCtx := client.Context{}.WithCodec(appCodec).WithHomeDir(home).WithOutput(out)
+
+			ctx := context.Background()
+			ctx = context.WithValue(ctx, client.ClientContextKey, &clientCtx)
+			ctx = context.WithValue(ctx, server.ServerContextKey, serverCtx)
+
+			if tc.expectError {
+				require.Error(t, cmd.ExecuteContext(ctx))
+			} else {
+				require.NoError(t, cmd.ExecuteContext(ctx))
+			}
+		})
 	}
 }